service_websocket/requirements.txt
websockets==8.1
PyJWT==1.7.1
sentry-sdk==0.14.4
loguru==0.5.3